Refactor : Test 코드와 Product 코드의 package 단계 맞춤
